cleaned up code into functions and made sure it was still working.
